requirements.txt
anyio==3.6.2
Appium-Python-Client==2.8.1
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
configuration.py==0.8.3
Deprecated==1.2.13
exceptiongroup==1.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.4
iniconfig==2.0.0
jproperties==2.1.1
jsonschema==4.17.3
outcome==1.2.0
packaging==23.0
Pillow==9.5.0
pluggy==1.0.0
pydantic==1.10.4
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.2.1
pytest-zebrunner==1.5.7
python-dotenv==0.21.1
PyYAML==5.4.1
requests==2.28.2
rfc3986==1.5.0
selenium==4.8.2
selenium-page-factory==2.5
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
urllib3==1.26.14
webcolors==1.13
wrapt==1.15.0
wsproto==1.2.0
